# perf(types): remove warning logic in binary doc construction