--- a/jina/types/document/__init__.py
+++ b/jina/types/document/__init__.py
@@ -291,20 +291,7 @@ class Document(ProtoTypeMixin, VersionedMixin):
                                 {k: document[k] for k in _remainder}
                             )
             elif isinstance(document, bytes):
-                # directly parsing from binary string gives large false-positive
-                # fortunately protobuf throws a warning when the parsing seems go wrong
-                # the context manager below converts this warning into exception and throw it
-                # properly
-                with warnings.catch_warnings():
-                    warnings.filterwarnings(
-                        'error', 'Unexpected end-group tag', category=RuntimeWarning
-                    )
-                    try:
-                        self._pb_body.ParseFromString(document)
-                    except RuntimeWarning as ex:
-                        raise BadDocType(
-                            f'fail to construct a document from {document}'
-                        ) from ex
+                self._pb_body.ParseFromString(document)
             elif isinstance(document, Document):
                 if copy:
                     self._pb_body.CopyFrom(document.proto)