background: https://github.com/jina-ai/jina/pull/3193#pullrequestreview-731999252

close #3193 

```
➜  jina git:(master) ✗ python toy39.py
get embedding ...	get embedding takes 4 seconds (4.49s)

➜  jina git:(perf-rm-warn-logic) ✗ python toy39.py
get embedding ...	get embedding takes 3 seconds (3.99s)
```