@alaeddine-13 2021-08-18
@hanxiao let me check how this changes affects detecting corrupted documents/dam before merging