@alaeddine-13 2021-08-18
> @hanxiao let me check how these changes affects detecting corrupted documents/dam before merging

looks fine